refactor: 오디오 플레이어를 audio-player.js로 분리 (ADR-034 PR1)
views-routing.js의 오디오 영역(showAudioPlayer/hideAudioBar/formatTime/
_teardownAudio/showAudioUnavailable/applyAudioShow + 재생 상태)을
js/app/audio-player.js로 추출. 2,389→2,195줄.

- route측 호출(renderChapter/renderPrologue)은 명시 ESM import로 결합 명시화
- 외부 호출자(search·settings·bookmark·app·sync)용 window.{hideAudioBar,
  applyAudioShow,getCurrentAudio} facade는 audio-player.js가 소유
- applyAudioShow→parsePath 상향 엣지만 window.parsePath 임시 경유(PR5 해소)
- index.html 스크립트 태그 + sw.js 셸 캐시 매니페스트 항목 추가

검증: tsc(main·worker) 0 error, 유닛 674 통과, 오디오 e2e 12 통과.
