+++ index.html
@@ -428,6 +428,7 @@
   <script type="module" src="/js/app/citations.js"></script>
   <script type="module" src="/js/app/parallels.js"></script>
   <script type="module" src="/js/app/tab-history.js"></script>
+  <script type="module" src="/js/app/audio-player.js"></script>
   <script type="module" src="/js/app/views-routing.js"></script>
   <script type="module" src="/js/app/tabbar.js"></script>
   <script type="module" src="/js/app.js"></script>
